# backend/apiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /follow:
        get:
            tags:
                - FollowService
            operationId: FollowService_ListFollowing
            parameters:
                - name: userId
                  in: query
                  description: '@gotags: json:"userId,omitempty,string"'
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListFollowingResponse'
        post:
            tags:
                - FollowService
            operationId: FollowService_AddFollow
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.AddFollowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.AddFollowResponse'
        delete:
            tags:
                - FollowService
            operationId: FollowService_RemoveFollow
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.RemoveFollowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.RemoveFollowResponse'
...
components:
    schemas:
...
        apiService.svapi.AddFollowRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
        apiService.svapi.AddFollowResponse:
            type: object
            properties: {}
...
        apiService.svapi.FollowUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        apiService.svapi.ListFollowingResponse:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.FollowUser'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
...
        apiService.svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        apiService.svapi.RemoveFollowRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
        apiService.svapi.RemoveFollowResponse:
            type: object
            properties: {}
...